feat: adding HATEOAS to Servicos

+++ src/main/java/br/com/fiap/barbertime/controller/ServicosController.java
@@ -1,12 +1,27 @@
 package br.com.fiap.barbertime.controller;
 
+import static org.springframework.http.HttpStatus.CREATED;
+import static org.springframework.http.HttpStatus.NOT_FOUND;
+import static org.springframework.http.HttpStatus.NO_CONTENT;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.data.web.PageableDefault;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.data.web.PagedResourcesAssembler;
+import org.springframework.hateoas.EntityModel;
+import org.springframework.hateoas.PagedModel;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
 import br.com.fiap.barbertime.model.Servicos;
@@ -18,10 +33,6 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 
-import static org.springframework.http.HttpStatus.CREATED;
-import static org.springframework.http.HttpStatus.NO_CONTENT;
-import static org.springframework.http.HttpStatus.NOT_FOUND;
-
 @RestController
 @RequestMapping("/servicos")
 @Tag(name = "Serviços", description = "Endpoints para operações relacionadas a serviços")
@@ -30,14 +41,19 @@ public class ServicosController {
     @Autowired
     private ServicosRepository servicosRepository;
 
+    @Autowired
+    PagedResourcesAssembler<Servicos> pagedResourcesAssembler;
+
     @GetMapping
     @Operation(
         summary = "Listar todos os serviços",
         description = "Retorna uma lista paginada de todos os serviços, ordenados pelo nome em ordem alfabética"
     )
-    public Page<Servicos> listarServicos(
+    public PagedModel<EntityModel<Servicos>> listarServicos(
             @Parameter(description = "Página solicitada (começando do 0)", example = "0") @PageableDefault(sort = "nome", direction = Direction.ASC) Pageable pageable) {
-        return servicosRepository.findAll(pageable);
+        
+        Page<Servicos> page = servicosRepository.findAll(pageable);
+        return pagedResourcesAssembler.toModel(page);
     }
 
     @GetMapping("/nome")
@@ -60,11 +76,13 @@ public class ServicosController {
         @ApiResponse(responseCode = "200", description = "Serviço encontrado"),
         @ApiResponse(responseCode = "404", description = "Serviço não encontrado", content = @Content)
     })
-    public ResponseEntity<Servicos> buscarPorId(
+    public EntityModel<Servicos> buscarPorId(
             @Parameter(description = "ID do serviço a ser buscado", example = "1") @PathVariable Long id) {
-        return servicosRepository.findById(id)
-                .map(ResponseEntity::ok)
-                .orElse(ResponseEntity.notFound().build());
+        
+        var servico = servicosRepository.findById(id)
+        .orElseThrow(() -> new ResponseStatusException(NOT_FOUND, "Serviço não encontrada"));
+
+        return servico.toEntityModel();
     }
 
     
@@ -95,7 +113,10 @@ public class ServicosController {
     })
     public void excluirServico(
             @Parameter(description = "ID do serviço a ser excluído", example = "1") @PathVariable Long id) {
-        verificarExistenciaServico(id);
+        if (!servicosRepository.existsById(id)) {
+            throw new ResponseStatusException(NOT_FOUND, "Barbearia não encontrada com o ID: " + id);
+        }
+        
         servicosRepository.deleteById(id);
     }
 
@@ -111,15 +132,14 @@ public class ServicosController {
     public Servicos atualizarServico(
             @Parameter(description = "ID do serviço a ser atualizado", example = "1") @PathVariable Long id,
             @Parameter(description = "Novos dados do serviço") @RequestBody Servicos servico) {
-        verificarExistenciaServico(id);
-        servico.setId(id);
-        return servicosRepository.save(servico);
+                if (!servicosRepository.existsById(id)) {
+                    throw new ResponseStatusException(NOT_FOUND, "Serviço não encontrada: ID " + id);
+                }
+        
+                servico.setId(id);
+                return servicosRepository.save(servico);
     }
 
-    private void verificarExistenciaServico(Long id) {
-        if (!servicosRepository.existsById(id)) {
-            throw new ResponseStatusException(NOT_FOUND, "Serviço não encontrado com o ID: " + id);
-        }
-    }
+
 
 }
